# Addressed performance degradation in Amazon.Runtime.ClientConfig