--- a/generator/.DevConfigs/20198df3-ff3a-44e2-a5d1-da81581bf0e1.json
+++ b/generator/.DevConfigs/20198df3-ff3a-44e2-a5d1-da81581bf0e1.json
@@ -1,0 +1,9 @@
+{
+    "core": {
+      "changeLogMessages": [
+        "Addressed performance degradation in Amazon.Runtime.ClientConfig ServiceURL"
+      ],
+      "type": "patch",
+      "updateMinimum": true
+    }
+  }